author-service/requirements.txt
fastapi
uvicorn
psycopg2-binary
gunicorn
kafka-python-ng
six

